+++ resources/views/TeamAssignment/viewTeamModal.blade.php
@@ -1,0 +1,38 @@
+
+    <div class="modal fade" id="viewTeamCompositionModal" tabindex="-1" role="dialog" aria-labelledby="addTeamLabel" aria-hidden="true">
+        <div class="modal-dialog modal-lg" role="document">
+            <div class="modal-content sectionDark">
+                <div class="modal-header"> 
+                    <small class="smCat">Team Composition</small>
+                </h4>
+                    <button type="button" class="close modalCloseButton" aria-label="Close">
+                        <span aria-hidden="true">&times;</span>
+                    </button>
+                </div>
+                <div class="modal-body modalBody">
+                    <div class="row">
+                        <div class="col-12">
+                            <div class="card text-center">
+                                <div class="card-header">
+                                    <div class="float-right">
+                                        <a data-collapse="#viewEmployees" class="btn btn-icon"><i class="ion ion-minus"></i></a>
+                                    </div>
+                                    <h5 class="text-center ml-5 cloneAppend">List of Employees</h5>
+                                </div>
+                                <div class="collapse show" id="viewEmployees">
+                                    <div class="card-body">
+                                        <div class="viewTeamCard row">
+
+                                        </div>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+                <div class="modal-footer">
+                    <button type="button" class="btn btn-primary waves-effect modalCloseButton">Close</button>
+                </div>
+            </div>
+        </div>
+    </div>